+++ go.mod
@@ -12,7 +12,7 @@ require (
 	github.com/golang/protobuf v1.3.3
 	github.com/googleapis/gnostic v0.4.1 // indirect
 	github.com/hashicorp/golang-lru v0.5.4
-	github.com/lyft/flyteidl v0.17.29
+	github.com/lyft/flyteidl v0.17.31-0.20200514231320-6542c9f82f75
 	github.com/lyft/flytestdlib v0.3.3
 	github.com/magiconair/properties v1.8.1
 	github.com/mitchellh/go-ps v1.0.0
