+++ go/tasks/pluginmachinery/flytek8s/container_helper.go
@@ -85,7 +85,7 @@ func ApplyResourceOverrides(ctx context.Context, resources v1.ResourceRequiremen
 }
 
 // Returns a K8s Container for the execution
-func ToK8sContainer(ctx context.Context, taskExecutionMetadata pluginsCore.TaskExecutionMetadata, taskContainer *core.Container,
+func ToK8sContainer(ctx context.Context, taskExecutionMetadata pluginsCore.TaskExecutionMetadata, taskContainer *core.Container, iFace *core.TypedInterface,
 	inputReader io.InputReader, outputPaths io.OutputFilePaths) (*v1.Container, error) {
 	modifiedCommand, err := utils.ReplaceTemplateCommandArgs(ctx, taskContainer.GetCommand(), inputReader, outputPaths)
 	if err != nil {
@@ -130,5 +130,8 @@ func ToK8sContainer(ctx context.Context, taskExecutionMetadata pluginsCore.TaskE
 		c.Resources = *res
 	}
 
+	if err := AddCoPilotToContainer(config.GetK8sPluginConfig().CoPilot, c, iFace, taskContainer.DataConfig); err != nil {
+		return nil, err
+	}
 	return c, nil
 }
